# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ad Seller System API",
    "description": "IAB OpenDirect 2.1 compliant seller agent for programmatic advertising. Supports product discovery, tiered pricing, proposal evaluation, multi-round negotiation, deal execution, order management, and change requests.",
    "contact": {
      "name": "IAB Tech Lab",
      "url": "https://iabtechlab.com/"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/change-requests": {
      "post": {
        "tags": [
          "Change Requests"
        ],
        "summary": "Create Change Request",
        "description": "Submit a change request for an existing order.\n\nValidates the change against the current order state, classifies\nseverity, and routes to approval if needed.",
        "operationId": "create_change_request_api_v1_change_requests_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChangeRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Change Requests"
        ],
        "summary": "List Change Requests",
        "description": "List change requests, optionally filtered by order or status.",
        "operationId": "list_change_requests_api_v1_change_requests_get",
        "parameters": [
          {
            "name": "order_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/change-requests/{cr_id}": {
      "get": {
        "tags": [
          "Change Requests"
        ],
        "summary": "Get Change Request",
        "description": "Get a change request by ID.",
        "operationId": "get_change_request_api_v1_change_requests__cr_id__get",
        "parameters": [
          {
            "name": "cr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cr Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/change-requests/{cr_id}/review": {
      "post": {
        "tags": [
          "Change Requests"
        ],
        "summary": "Review Change Request",
        "description": "Approve or reject a pending change request.",
        "operationId": "review_change_request_api_v1_change_requests__cr_id__review_post",
        "parameters": [
          {
            "name": "cr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cr Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewChangeRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/change-requests/{cr_id}/apply": {
      "post": {
        "tags": [
          "Change Requests"
        ],
        "summary": "Apply Change Request",
        "description": "Apply an approved change request to the order.\n\nUpdates the order with the proposed values from the change request.",
        "operationId": "apply_change_request_api_v1_change_requests__cr_id__apply_post",
        "parameters": [
          {
            "name": "cr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cr Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Api-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateChangeRequestModel": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "change_type": {
            "type": "string",
            "title": "Change Type"
          },
          "diffs": {
            "items": {
              "$ref": "#/components/schemas/FieldDiffModel"
            },
            "type": "array",
            "title": "Diffs",
            "default": []
          },
          "proposed_values": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Values"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "requested_by": {
            "type": "string",
            "title": "Requested By",
            "default": "system"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "change_type"
        ],
        "title": "CreateChangeRequestModel",
        "description": "Request to create a change request for an order."
      },
...
      "FieldDiffModel": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "old_value": {
            "title": "Old Value"
          },
          "new_value": {
            "title": "New Value"
          }
        },
        "type": "object",
        "required": [
          "field"
        ],
        "title": "FieldDiffModel"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewChangeRequestModel": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "decided_by": {
            "type": "string",
            "title": "Decided By",
            "default": "system"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ReviewChangeRequestModel",
        "description": "Approve or reject a change request."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}